+++ include/World.h
@@ -34,23 +34,23 @@ class World {
      */
     void destroy();
 
-    void addRoom(Room& room);
+    void addRoom(Room* room);
     unsigned long sizeRoom();
     Room& getRoom(unsigned long index);
 
-    void addSprite(SpriteSequence& sprite);
+    void addSprite(SpriteSequence* sprite);
     unsigned long sizeSprite();
     SpriteSequence& getSprite(unsigned long index);
 
-    void addEntity(Entity& entity);
+    void addEntity(Entity* entity);
     unsigned long sizeEntity();
     Entity& getEntity(unsigned long index);
 
-    void addSkeletalModel(SkeletalModel& model);
+    void addSkeletalModel(SkeletalModel* model);
     unsigned long sizeSkeletalModel();
     SkeletalModel& getSkeletalModel(unsigned long index);
 
-    void addStaticMesh(StaticMesh& model);
+    void addStaticMesh(StaticMesh* model);
     unsigned long sizeStaticMesh();
     StaticMesh& getStaticMesh(unsigned long index);